% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_generator_functions.R
\name{fgen_force_length}
\alias{fgen_force_length}
\title{Force Length Relationship}
\usage{
fgen_force_length(
  current_distance,
  start_force = 0.8,
  end_force = 0,
  threshold = 0.9,
  push_off_distance = 0.4,
  ...
)
}
\arguments{
\item{current_distance}{Numeric vector}

\item{start_force}{Numeric value. Indicates starting percentage. Allowed range from 0.5 to 1. Default is 0.8}

\item{end_force}{Numeric value. Indicates ending percentage. Allowed range from 0.5 to 1. Default is 0}

\item{threshold}{Numeric value. Indicates threshold where the line starts to break. Use values from 0 to 1, where
1 is equal to \code{push_off_distance} position. Allowed range from 0.8 to 1. Default is 0.9}

\item{push_off_distance}{Numeric value. Indicates the push off distance/length. Default is 0.4.}

\item{...}{Used to allow different parameters to be passes without error}
}
\value{
Numeric vector with values from 0 to 1, indicating percentage of Force at particular \code{current_distance}
}
\description{
Force Length Relationship
}
\examples{
x <- seq(0, 0.5, length.out = 1000)

y1 <- fgen_force_length(
  current_distance = x,
  start_force = 0.6,
  threshold = 0.90,
  end = 0
)

plot(x, y1, "l")

y2 <- fgen_force_length(
  current_distance = x,
  start_force = 0.8,
  threshold = 0.95,
  end = 0
)

lines(x, y2, col = "red")
}
