% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_generator_functions.R
\name{fgen_force_time}
\alias{fgen_force_time}
\title{Force Time Relationship}
\usage{
fgen_force_time(
  current_time,
  initial_force = 0,
  max_force = 3000,
  time_to_max_force = 0.3,
  ...
)
}
\arguments{
\item{current_time}{Numeric vector}

\item{initial_force}{Numeric value. Default is 0}

\item{max_force}{Numeric value. Default is 1}

\item{time_to_max_force}{Numeric value. Time it takes to reach \code{max_force} from 0.
This means that is \code{initial_force} is higher than 0, it will take less time
to reack \code{max_force}. Default is 0.3s.}

\item{...}{Used to allow different parameters to be passes without error}
}
\value{
Numeric vector. Force level at \code{current_time}
}
\description{
Force Time Relationship
}
\examples{
x <- seq(0, 1, length.out = 1000)

y1 <- fgen_force_time(
  current_time = x,
  initial_force = 0,
  max_force = 1000,
  time_to_max_force = 0.5
)

plot(x, y1, "l")

y2 <- fgen_force_time(
  current_time = x,
  initial_force = 200,
  max_force = 1000,
  time_to_max_force = 0.5
)

lines(x, y2, col = "red")

y3 <- fgen_force_time(
  current_time = x,
  initial_force = 200,
  max_force = 1000,
  time_to_max_force = 0.3
)

lines(x, y3, col = "blue")
}
