% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_generator_functions.R
\name{fgen_force_velocity}
\alias{fgen_force_velocity}
\title{Force Velocity Relationhsip}
\usage{
fgen_force_velocity(current_velocity, max_force = 3000, max_velocity = 4, ...)
}
\arguments{
\item{current_velocity}{Numeric vector}

\item{max_force}{Numeric value. Default is 3000 Newtons}

\item{max_velocity}{Numeric value. Default is 4 m/s}

\item{...}{Used to allow different parameters to be passes without error}
}
\value{
Viscous force at \code{current_velocity}
}
\description{
Force Velocity Relationhsip
}
\examples{
x <- seq(0, 3, length.out = 1000)

y1 <- fgen_force_velocity(
  current_velocity = x,
  max_velocity = 3,
  max_force = 1200
)

plot(x, y1, "l")

y2 <- fgen_force_velocity(
  current_velocity = x,
  max_velocity = 3,
  max_force = 1000
)

lines(x, y2, col = "red")
}
