% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgen.R
\name{fgen_get_activation}
\alias{fgen_get_activation}
\title{Get Force Generator Activation Level}
\usage{
fgen_get_activation(
  current_time,
  initial_activation = 0,
  time_to_max_activation = 0.3
)
}
\arguments{
\item{current_time}{Numeric vector}

\item{initial_activation}{Numeric vector between 0 and 1. Default is 0}

\item{time_to_max_activation}{Numeric vector. Time it takes to reach maximal activation from 0. Default is 0.3s.}
}
\value{
Numeric vector. Activation level at \code{current_time}
}
\description{
Due to the Force Generator Force-Time characteristic, it takes time to activate the Force Generator.
    \code{fgen_get_activation} is a function that represents the activation level at \code{current_time}
}
\examples{
x <- seq(0, 1, length.out = 1000)

y1 <- fgen_get_activation(
  current_time = x,
  initial_activation = 0,
  time_to_max_activation = 0.5
)

plot(x, y1, "l")

y2 <- fgen_get_activation(
  current_time = x,
  initial_activation = 0.2,
  time_to_max_activation = 0.5
)

lines(x, y2, col = "red")

y3 <- fgen_get_activation(
  current_time = x,
  initial_activation = 0,
  time_to_max_activation = 0.4
)

lines(x, y3, col = "blue")
}
