% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgen.R
\name{fgen_get_force_percentage}
\alias{fgen_get_force_percentage}
\title{Get Force Generator Force Percentage}
\usage{
fgen_get_force_percentage(
  current_distance = 0,
  push_off_distance = 0.4,
  decline_rate = 1.05,
  peak_location = -push_off_distance * 0.15
)
}
\arguments{
\item{current_distance}{Numeric vector}

\item{push_off_distance}{Numeric vector. Default is 0.4}

\item{decline_rate}{Numeric vector. Indicates how quickly does the force percentage drops before \code{peak_location}. Default is 1.05}

\item{peak_location}{Numeric vector. Indicates location of the peak force (100%). It is represented with the distance from the take-off position.
Default is \code{-push_off_distance * 0.15}}
}
\value{
Numeric vector with values from 0 to 1, indicating Force Percentage at particular \code{current_distance}
}
\description{
Due to the Force Generator Force-Length characteristic, maximal force is unable to be expressed across whole Push-off Distance.
   \code{fgen_get_force_percentage} is a function that represents the force percentage at \code{current_distance}
}
\examples{
x <- seq(0, 0.5, length.out = 1000)

y1 <- fgen_get_force_percentage(
  current_distance = x,
  push_off_distance = 0.4,
  decline_rate = 1.1,
  peak_location = -0.1
)

plot(x, y1, "l")

y2 <- fgen_get_force_percentage(
  current_distance = x,
  push_off_distance = 0.4,
  decline_rate = 1.5,
  peak_location = -0.2
)

lines(x, y2, col = "red")

y3 <- fgen_get_force_percentage(
  current_distance = x,
  push_off_distance = 0.5,
  decline_rate = 1.5,
  peak_location = -0.2
)

lines(x, y3, col = "blue")
}
