% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgen.R
\name{fgen_get_velocity}
\alias{fgen_get_velocity}
\title{Get Force-Velocity Velocity of the Force Generator}
\usage{
fgen_get_velocity(external_force, max_force = 3000, max_velocity = 4)
}
\arguments{
\item{external_force}{Numeric vector. External force is in Newtons}

\item{max_force}{Numeric vector. Maximal force in Newton that Force Generator can generate}

\item{max_velocity}{Numeric vector. Maximal velocity that Force Generator can achieve in unconstrained conditions}
}
\value{
Numeric vector of estimated maximal velocity reached
}
\description{
Based on the parameters of the Force Generator: \code{max_force} and \code{max_velocity}
    get maximal velocity that can be reached at particular \code{external_resistance}
}
\examples{
fgen_get_velocity(0, 3000, 4)
fgen_get_velocity(3000, 3000, 4)
fgen_get_velocity(1500, 3000, 4)
}
