% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgen.R
\name{fgen_get_viscous_force}
\alias{fgen_get_viscous_force}
\title{Get Force Generator Viscous Force}
\usage{
fgen_get_viscous_force(current_velocity, max_force = 3000, max_velocity = 4)
}
\arguments{
\item{current_velocity}{Numeric vector}

\item{max_force}{Numeric value. Default is 3000 Newtons}

\item{max_velocity}{Numeric value. Default is 4 m/s}
}
\value{
Viscous force at \code{current_velocity}
}
\description{
Due to the Force Generator Force-Velocity characteristic, it velocity produces breaking or viscous force.
    \code{fgen_get_viscous_force} is a function that represents the viscous force at \code{current_velocity}
}
\examples{
x <- seq(0, 3, length.out = 1000)

y1 <- fgen_get_viscous_force(
  current_velocity = x,
  max_velocity = 3,
  max_force = 1200
)

plot(x, y1, "l")

y2 <- fgen_get_viscous_force(
  current_velocity = x,
  max_velocity = 3,
  max_force = 1000
)

lines(x, y2, col = "red")
}
