% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aerial_time}
\alias{get_aerial_time}
\title{Get Aerial Time}
\usage{
get_aerial_time(height, gravity_const = 9.81)
}
\arguments{
\item{height}{Numeric vector. Jump height in meters}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector. Aerial time in seconds
}
\description{
\code{get_aerial_time} returns aerial time from the jump \code{height}
}
\examples{
get_aerial_time(0.4)
}
