% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samozino.R
\name{get_all_samozino_profiles}
\alias{get_all_samozino_profiles}
\title{Get All Samozino profiles}
\usage{
get_all_samozino_profiles(profile_data)
}
\arguments{
\item{profile_data}{Data frame returned from \code{\link{vj_profile}} function}
}
\value{
List with two elements: list of profiles and long data frame
}
\description{
\code{get_all_samozino_profiles} returns theoretical and practical Samozino profiles using \code{profile_data}
returned from the \code{\link{vj_profile}} and \code{\link{get_samozino_optimal_profile}} functions
}
\examples{
fv_profile <- vj_profile(mass = 75)
all_samozino_profiles <- get_all_samozino_profiles(fv_profile)
all_samozino_profiles$data_frame
}
\references{
Samozino, Pierre. ‘A Simple Method for Measuring Lower Limb Force, Velocity and Power Capabilities During Jumping’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 65–96. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_4.

    ———. ‘Optimal Force-Velocity Profile in Ballistic Push-off: Measurement and Relationship with Performance’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 97–119. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_5.

    Samozino, Pierre, Jean-Benoît Morin, Frédérique Hintzy, and Alain Belli. ‘Jumping Ability: A Theoretical Integrative Approach’. Journal of Theoretical Biology 264, no. 1 (May 2010): 11–18. https://doi.org/10.1016/j.jtbi.2010.01.021.

    Samozino, Pierre, Enrico Rejc, Pietro Enrico Di Prampero, Alain Belli, and Jean-Benoît Morin. ‘Optimal Force–Velocity Profile in Ballistic Movements—Altius’: Medicine & Science in Sports & Exercise 44, no. 2 (February 2012): 313–22. https://doi.org/10.1249/MSS.0b013e31822d757a.
}
