% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_FV_profile}
\alias{get_FV_profile}
\title{Get Force-Velocity profile}
\usage{
get_FV_profile(
  profile_data,
  force = "mean_GRF_over_distance",
  velocity = "mean_velocity"
)
}
\arguments{
\item{profile_data}{Data frame returned from \code{\link{vj_profile}} function}

\item{force}{Character string. Name of the column in the \code{profile_data}. Default is "mean_GRF_over_distance"}

\item{velocity}{Character string. Name of the column in the \code{profile_data}. Default is "mean_velocity"}
}
\value{
List with calculated \code{F0}, \code{V0}, \code{Pmax}, and \code{Sfv}
}
\description{
Applied simple linear regression model using \code{force} and \code{velocity} columns from the \code{profile_data} and
    returns calculated \code{F0}, \code{V0}, \code{Pmax}, and \code{Sfv}
}
\examples{
fv_profile <- vj_profile(mass = 75)
get_FV_profile(fv_profile)

}
