% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_height}
\alias{get_height}
\title{Get Height}
\usage{
get_height(take_off_velocity, gravity_const = 9.81)
}
\arguments{
\item{take_off_velocity}{Numeric vector}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector
}
\description{
Function calculates height reached using the ballistic equation
}
\examples{
get_height(2.75)
}
