% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_height_from_aerial_time}
\alias{get_height_from_aerial_time}
\title{Get Height from Aerial Time}
\usage{
get_height_from_aerial_time(aerial_time, gravity_const = 9.81)
}
\arguments{
\item{aerial_time}{Numeric vector. Aerial time in seconds}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector. Jump height in meters
}
\description{
\code{get_height_from_aerial_time} returns jump height using \code{aerial_time}
}
\examples{
get_height_from_aerial_time (0.7)
}
