% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_impuse}
\alias{get_impuse}
\title{Get Impulse}
\usage{
get_impuse(mass, take_off_velocity)
}
\arguments{
\item{mass}{Numeric vector}

\item{take_off_velocity}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Get Impulse
}
\examples{
get_impuse(75, 2)
}
