% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_max_power}
\alias{get_max_power}
\title{Get Max Power}
\usage{
get_max_power(max_force, max_velocity)
}
\arguments{
\item{max_force}{Numeric vector}

\item{max_velocity}{Numeric vector.}
}
\value{
Numeric vector
}
\description{
Function \code{get_max_power} calculates maximal power assuming linear relationship between
    \code{max_force} and \code{max_velocity}
}
\examples{
get_max_power(3000, 4)
}
