% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mean_force_over_time}
\alias{get_mean_force_over_time}
\title{Get Mean Force over Time}
\usage{
get_mean_force_over_time(
  mass,
  weight = mass * 9.81,
  take_off_velocity,
  time_taken
)
}
\arguments{
\item{mass}{Numeric vector}

\item{weight}{Numeric vector. Default \code{mass} * 9.81}

\item{take_off_velocity}{Numeric vector}

\item{time_taken}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Function \code{get_mean_force_over_time} calculates mean force over push-off time using impulse equation
}
\examples{
get_mean_force_over_time(100, 100*9.81, 4, 0.3)
}
