% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mean_power}
\alias{get_mean_power}
\title{Get Mean Power}
\usage{
get_mean_power(
  mass,
  weight = mass * 9.81,
  take_off_velocity,
  push_off_distance,
  time_taken
)
}
\arguments{
\item{mass}{Numeric vector}

\item{weight}{Numeric vector. Default \code{mass} * 9.81}

\item{take_off_velocity}{Numeric vector}

\item{push_off_distance}{Numeric vector}

\item{time_taken}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Function \code{get_mean_power} calculates the vertical jump mean power using kinetic and potential enery
}
\examples{
get_mean_power(100, 100 * 9.81, 4, 0.4, 0.3)
}
