% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_power}
\alias{get_power}
\title{Get Power}
\usage{
get_power(max_force, max_velocity)
}
\arguments{
\item{max_force}{Numeric vector}

\item{max_velocity}{Numeric vector. Default 9.81}
}
\value{
Numeric vector
}
\description{
Function calculates maximal power assuming linear relationship between
    \code{max_force} and \code{max_velocity}
}
\examples{
get_power(3000, 4)
}
