% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probing.R
\name{get_probing_data}
\alias{get_probing_data}
\title{Get Probing Data}
\usage{
get_probing_data(
  args_list,
  probe_func = function(...) { },
  change_ratio = seq(0.8, 1.2, length.out = 100),
  aggregate = "raw",
  ...
)
}
\arguments{
\item{args_list}{Named list with initial arguments to be passed to \code{probe_func}}

\item{probe_func}{Function that returns a list of outputs}

\item{change_ratio}{Numeric vector indicating the probing initial arguments in \code{args_list}}

\item{aggregate}{String. Indicates how should output from \code{probe_func} be aggregated.
Default is "raw". Other options involve "diff" and "ratio". To calculate "diff" and
"ratio" aggregate, \code{probe_func} output using the initial argument values in
\code{args_list} is used}

\item{...}{Extra arguments for \code{probe_func} that are not probed using \code{change_ratio}}
}
\value{
Data frame
}
\description{
Get Probing Data
}
\examples{
require(ggplot2)
fgen_probe_data <- get_probing_data(
  args_list = list(current_time = 0.3, current_distance = 0.1, current_velocity = 0),
  probe_func = function(...) {
    fgen_get_output(...)
  }, aggregate = "raw",
  mass = 70, max_force = 5000
)

ggplot(
  fgen_probe_data,
  aes(
    x = change_ratio,
    y = kinetics.ground_reaction_force,
    color = probing
  )
) +
  geom_line()
}
