% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samozino.R
\name{get_samozino_jump_metrics}
\alias{get_samozino_jump_metrics}
\title{Get Samozino Jump Metrics}
\usage{
get_samozino_jump_metrics(
  mass = 75,
  push_off_distance = 0.4,
  height = 0.5,
  gravity_const = 9.81
)
}
\arguments{
\item{mass}{Numeric vector. Default 75}

\item{push_off_distance}{Numeric vector. Default 0.4}

\item{height}{Numeric vector. Jump height. Default 0.5}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
List with \code{mass}, \code{push_off_distance}, \code{height}, \code{gravity_const}, \code{mean_GRF_over_distance},
        \code{mean_velocity}, \code{take_off_velocity}, and \code{mean_power}
}
\description{
\code{get_samozino_jump_metrics} return mean force, mean velocity and mean power using simple method explained in
the references.
}
\examples{
get_samozino_jump_metrics(
  mass = 75,
  push_off_distance = 0.43,
  height = 0.45
)
}
\references{
Samozino, Pierre. ‘A Simple Method for Measuring Lower Limb Force, Velocity and Power Capabilities During Jumping’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 65–96. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_4.

    ———. ‘Optimal Force-Velocity Profile in Ballistic Push-off: Measurement and Relationship with Performance’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 97–119. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_5.

    Samozino, Pierre, Jean-Benoît Morin, Frédérique Hintzy, and Alain Belli. ‘Jumping Ability: A Theoretical Integrative Approach’. Journal of Theoretical Biology 264, no. 1 (May 2010): 11–18. https://doi.org/10.1016/j.jtbi.2010.01.021.

    Samozino, Pierre, Enrico Rejc, Pietro Enrico Di Prampero, Alain Belli, and Jean-Benoît Morin. ‘Optimal Force–Velocity Profile in Ballistic Movements—Altius’: Medicine & Science in Sports & Exercise 44, no. 2 (February 2012): 313–22. https://doi.org/10.1249/MSS.0b013e31822d757a.
}
