% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samozino.R
\name{get_samozino_optimal_profile}
\alias{get_samozino_optimal_profile}
\title{Get Samozino's Optimal Profile}
\usage{
get_samozino_optimal_profile(
  F0 = 3000,
  V0 = 4,
  bodyweight = 75,
  push_off_distance = 0.4,
  gravity_const = 9.81
)
}
\arguments{
\item{F0}{Numeric value. Default 3000}

\item{V0}{Numeric value. Default 4}

\item{bodyweight}{Numeric value. Default 75}

\item{push_off_distance}{Numeric value. Default 0.4}

\item{gravity_const}{Numeric value. Default 9.81}
}
\value{
List with list and data frame
}
\description{
\code{get_samozino_optimal_profile} finds optimal Force-Velocity profile that maximizes vertical jump height based on the model of
 Samozino \emph{et al.}. According to this model, we are looking for Force-Velocity profile that maintains the max power, but changes
 \code{F0} and \code{V0} so it maximized the take-off velocity calculated using \code{\link{get_samozino_take_off_velocity}}.
}
\examples{
get_samozino_optimal_profile(F0 = 2500, V0 = 3.7, bodyweight = 85, push_off_distance = 0.42)
}
\references{
Samozino, Pierre. ‘A Simple Method for Measuring Lower Limb Force, Velocity and Power Capabilities During Jumping’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 65–96. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_4.

    ———. ‘Optimal Force-Velocity Profile in Ballistic Push-off: Measurement and Relationship with Performance’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 97–119. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_5.

    Samozino, Pierre, Jean-Benoît Morin, Frédérique Hintzy, and Alain Belli. ‘Jumping Ability: A Theoretical Integrative Approach’. Journal of Theoretical Biology 264, no. 1 (May 2010): 11–18. https://doi.org/10.1016/j.jtbi.2010.01.021.

    Samozino, Pierre, Enrico Rejc, Pietro Enrico Di Prampero, Alain Belli, and Jean-Benoît Morin. ‘Optimal Force–Velocity Profile in Ballistic Movements—Altius’: Medicine & Science in Sports & Exercise 44, no. 2 (February 2012): 313–22. https://doi.org/10.1249/MSS.0b013e31822d757a.
}
