% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samozino.R
\name{get_samozino_take_off_velocity}
\alias{get_samozino_take_off_velocity}
\title{Get Samozino Take-Off Velocity}
\usage{
get_samozino_take_off_velocity(
  F0 = 3000,
  V0 = 4,
  bodyweight = 75,
  push_off_distance = 0.4,
  gravity_const = 9.81
)
}
\arguments{
\item{F0}{Numeric vector. Default 3000}

\item{V0}{Numeric vector. Default 4}

\item{bodyweight}{Numeric vector. Default 75}

\item{push_off_distance}{Numeric vector. Default 0.4}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector representing maximal take-off velocity
}
\description{
\code{get_samozino_take_off_velocity} returns predicted maximal take off velocity that could be achieve based on the Samozino \emph{et al.}
    model which uses vertical jump profile \code{F0} and \code{V0}.
}
\examples{
get_samozino_take_off_velocity(F0 = 2500, V0 = 3.7, bodyweight = 85, push_off_distance = 0.42)
}
\references{
Samozino, Pierre. ‘A Simple Method for Measuring Lower Limb Force, Velocity and Power Capabilities During Jumping’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 65–96. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_4.

    ———. ‘Optimal Force-Velocity Profile in Ballistic Push-off: Measurement and Relationship with Performance’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 97–119. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_5.

    Samozino, Pierre, Jean-Benoît Morin, Frédérique Hintzy, and Alain Belli. ‘Jumping Ability: A Theoretical Integrative Approach’. Journal of Theoretical Biology 264, no. 1 (May 2010): 11–18. https://doi.org/10.1016/j.jtbi.2010.01.021.

    Samozino, Pierre, Enrico Rejc, Pietro Enrico Di Prampero, Alain Belli, and Jean-Benoît Morin. ‘Optimal Force–Velocity Profile in Ballistic Movements—Altius’: Medicine & Science in Sports & Exercise 44, no. 2 (February 2012): 313–22. https://doi.org/10.1249/MSS.0b013e31822d757a.
}
