% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_optimal_profile.R
\name{get_simple_optimal_profile}
\alias{get_simple_optimal_profile}
\title{Get Simple Optimal Profile}
\usage{
get_simple_optimal_profile(L0, TOV0, bodyweight, gravity_const = 9.81)
}
\arguments{
\item{L0}{Numeric vector}

\item{TOV0}{Numeric vector}

\item{bodyweight}{Numeric vector}

\item{gravity_const}{Numeric vector. Default is 9.81}
}
\value{
List
}
\description{
\code{get_simple_optimal_profile} calculates optimal profile using Load~TOV profile. This is done by
    assuming same 'surface' under the profile and getting the profile tha maximize TOV at bodyweight
}
\examples{
get_simple_optimal_profile(L0 = 250, TOV0 = 3, bodyweight = 75)
}
