% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_optimal_profile.R
\name{get_simple_profile}
\alias{get_simple_profile}
\title{Get Simple Profile}
\usage{
get_simple_profile(profile_data)
}
\arguments{
\item{profile_data}{Data frame returned from \code{\link{vj_profile}} function}
}
\value{
List returned by \code{\link{get_simple_optimal_profile}} function
}
\description{
\code{get_simple_profile} take the \code{rofile_data} and performs Load~TOV profile and returns
    the optimal profile using \code{\link{get_simple_optimal_profile}} function
}
\examples{
fv_profile <- vj_profile(mass = 75)
get_simple_profile(fv_profile)
}
