% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_optimal_profile.R
\name{get_simple_take_off_velocity}
\alias{get_simple_take_off_velocity}
\title{Get Simple Take-Off Velocity}
\usage{
get_simple_take_off_velocity(L0, TOV0, bodyweight)
}
\arguments{
\item{L0}{Numeric vector}

\item{TOV0}{Numeric vector}

\item{bodyweight}{Numeric vector}
}
\value{
Numeric vector
}
\description{
\code{get_simple_take_off_velocity} return predicted take-off velocity at bodyweight load
    using Load~Take-Off Velocity profile
}
\examples{
get_simple_take_off_velocity(L0 = 250, TOV0 = 3, bodyweight = 75)
}
