% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_slope}
\alias{get_slope}
\title{Get Force Velocity Profile Slop}
\usage{
get_slope(max_force, max_velocity)
}
\arguments{
\item{max_force}{Numeric vector}

\item{max_velocity}{Numeric vector.}
}
\value{
Numeric vector
}
\description{
Function \code{get_slope} calculates force-velocity profile slope assuming linear relationship
   between \code{max_force} and \code{max_velocity}
}
\examples{
get_slope(3000, 4)
}
