% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_take_off_velocity}
\alias{get_take_off_velocity}
\title{Get Take-off Velocity}
\usage{
get_take_off_velocity(
  mean_force = 3000,
  mass = 75,
  push_off_distance = 0.4,
  gravity_const = 9.81
)
}
\arguments{
\item{mean_force}{Numeric vector. Default 3000}

\item{mass}{Numeric vector. Default 75}

\item{push_off_distance}{Numeric vector. Default 0.4}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector
}
\description{
\code{get_take_off_velocity} returns take off velocity when \code{mean_force} is applied to the object
of \code{mass} over \code{push_off_distance} in vertical direction, assuming zero starting velocity
}
\examples{
get_take_off_velocity(
  mean_force = 2000,
  mass= 85,
  push_off_distance = 0.4
)
}
