% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_work}
\alias{get_work}
\title{Get Work Done}
\usage{
get_work(mass, weight = mass * 9.81, take_off_velocity, push_off_distance)
}
\arguments{
\item{mass}{Numeric vector}

\item{weight}{Numeric vector. Default \code{mass} * 9.81}

\item{take_off_velocity}{Numeric vector}

\item{push_off_distance}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Function \code{get_work} calculates the work needed for the vertical jump using
    kinetic and potential energy
}
\examples{
get_work(100, 100 * 9.81, 4, 0.4)
}
