% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samozino.R
\name{make_samozino_profile}
\alias{make_samozino_profile}
\title{Make Samozino Profile}
\usage{
make_samozino_profile(
  bodyweight,
  push_off_distance,
  external_load,
  height,
  aerial_time,
  gravity_const = 9.81,
  plot = TRUE
)
}
\arguments{
\item{bodyweight}{Numeric value}

\item{push_off_distance}{Numeric value}

\item{external_load}{Numeric vector}

\item{height}{Numeric vector}

\item{aerial_time}{Numeric vector}

\item{gravity_const}{Numeric vector. Default 9.81}

\item{plot}{TRUE/FALSE. Default is TRUE}
}
\value{
List of Samozino profile parameters
}
\description{
\code{make_samozino_profile} generates Samozino profile from squat jump data. When both \code{aerial_time} and
    \code{height} are forwarded to the function, only \code{height} is used to estimate mean GRF.
}
\examples{
require(tidyverse)

data("testing_data")

with(
  filter(testing_data, athlete == "Jack"),
  make_samozino_profile(
    bodyweight = bodyweight,
    push_off_distance = push_off_distance,
    external_load = external_load,
    aerial_time = aerial_time,
    plot = TRUE
  )
)
}
\references{
Samozino, Pierre. ‘A Simple Method for Measuring Lower Limb Force, Velocity and Power Capabilities During Jumping’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 65–96. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_4.

    ———. ‘Optimal Force-Velocity Profile in Ballistic Push-off: Measurement and Relationship with Performance’. In Biomechanics of Training and Testing, edited by Jean-Benoit Morin and Pierre Samozino, 97–119. Cham: Springer International Publishing, 2018. https://doi.org/10.1007/978-3-319-05633-3_5.

    Samozino, Pierre, Jean-Benoît Morin, Frédérique Hintzy, and Alain Belli. ‘Jumping Ability: A Theoretical Integrative Approach’. Journal of Theoretical Biology 264, no. 1 (May 2010): 11–18. https://doi.org/10.1016/j.jtbi.2010.01.021.

    Samozino, Pierre, Enrico Rejc, Pietro Enrico Di Prampero, Alain Belli, and Jean-Benoît Morin. ‘Optimal Force–Velocity Profile in Ballistic Movements—Altius’: Medicine & Science in Sports & Exercise 44, no. 2 (February 2012): 313–22. https://doi.org/10.1249/MSS.0b013e31822d757a.
}
