% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samozino.R
\name{probe_samozino_take_off_velocity}
\alias{probe_samozino_take_off_velocity}
\title{Probe Samozino Take-off Velocity}
\usage{
probe_samozino_take_off_velocity(
  F0 = 3000,
  V0 = 4,
  bodyweight = 75,
  push_off_distance = 0.4,
  gravity_const = 9.81,
  change_ratio = seq(0.9, 1.1, length.out = 3),
  aggregate = "raw"
)
}
\arguments{
\item{F0}{Numeric vector. Default 3000}

\item{V0}{Numeric vector. Default 4}

\item{bodyweight}{Numeric vector. Default 75}

\item{push_off_distance}{Numeric vector. Default 0.4}

\item{gravity_const}{Numeric vector. Default 9.81}

\item{change_ratio}{Numeric vector indicating probing change ratios}

\item{aggregate}{How should \code{\link{get_samozino_take_off_velocity}} output be aggregated?
Default is "raw". Other options involve "ratio" and "diff" which use initial
output values}
}
\value{
Probing data frame
}
\description{
\code{probe_samozino_take_off_velocity} probes results of the \code{\link{get_samozino_take_off_velocity}} function by varying
\code{F0}, \code{V0}, and \code{bodyweight} parameters
}
\examples{
require(ggplot2)

samozino_probe_data <- probe_samozino_take_off_velocity(
  F0 = 3000,
  V0 = 3.5,
  push_off_distance = 0.4,
  bodyweight = 75,
  change_ratio = seq(0.8, 1.2, length.out = 1001)
)

ggplot(
  samozino_probe_data,
  aes(
    x = change_ratio,
    y = take_off_velocity,
    color = probing
  )
) +
  geom_line()
}
