% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_optimal_profile.R
\name{probe_simple_take_off_velocity}
\alias{probe_simple_take_off_velocity}
\title{Probe Simple Take-off Velocity}
\usage{
probe_simple_take_off_velocity(
  L0,
  TOV0,
  bodyweight,
  change_ratio = seq(0.9, 1.1, length.out = 3),
  aggregate = "raw"
)
}
\arguments{
\item{L0}{Numeric vector}

\item{TOV0}{Numeric vector}

\item{bodyweight}{Numeric vector}

\item{change_ratio}{Numeric vector indicating probing change ratios}

\item{aggregate}{How should \code{\link{probe_simple_take_off_velocity}} output be aggregated?
Default is "raw". Other options involve "ratio" and "diff" which use initial
output values}
}
\value{
Probing data frame
}
\description{
\code{probe_simple_take_off_velocity} probes results of the \code{\link{get_simple_take_off_velocity}} function by varying
\code{L0}, \code{TOV0}, and \code{bodyweight} parameters
}
\examples{
require(ggplot2)

simple_probe_data <- probe_simple_take_off_velocity(
  L0 = 250,
  TOV0 = 3,
  bodyweight = 75,
  change_ratio = seq(0.8, 1.2, length.out = 1001)
)

ggplot(
  simple_probe_data,
  aes(
    x = change_ratio,
    y = take_off_velocity,
    color = probing
  )
) +
geom_line()
}
