% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgen.R
\name{fgen_get_force_percentage}
\alias{fgen_get_force_percentage}
\title{Get Force Generator Force Percentage}
\usage{
fgen_get_force_percentage(push_off_perc, start_perc = 0.8, threshold = 0.9)
}
\arguments{
\item{push_off_perc}{Numeric vector. Indicates percentage of the Push-off Distance. Values needs to be positive}

\item{start_perc}{Numeric vector. Indicates starting percentage. Allowed range from 0.5 to 1. Default is 0.8}

\item{threshold}{Numeric vector. Indicates threshold where the line starts to break. Use values from 0.8 to 1.
Default is 0.9}
}
\value{
Numeric vector with values from 0 to 1, indicating Force Percentage at particular \code{push_off_perc}
}
\description{
Due to the Force Generator Force-Length characteristic, maximal force is unable to be expressed across whole Push-off Distance.
   \code{fgen_get_force_percentage} is a function that represents the force percentage at \code{push_off_perc}
}
\examples{
x <- seq(0, 1.1, length.out = 1000)

y1 <- fgen_get_force_percentage(
  push_off_perc = x,
  start_perc = 0.6,
  threshold = 0.90
)

plot(x, y1, "l")

y2 <- fgen_get_force_percentage(
  push_off_perc = x,
  start_perc = 0.8,
  threshold = 0.95
)

lines(x, y2, col = "red")
}
