% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgen.R
\name{fgen_get_output}
\alias{fgen_get_output}
\title{Get Instanteous Output Of The Force Generator}
\usage{
fgen_get_output(
  current_time = 0,
  current_distance = 0,
  current_velocity = 0,
  mass = 75,
  weight = mass * 9.81,
  push_off_distance = 0.4,
  max_force = 3000,
  max_velocity = 4,
  start_perc = 0.8,
  threshold = 0.9,
  time_to_max_activation = 0.3
)
}
\arguments{
\item{current_time}{Numeric vector. Default is 0}

\item{current_distance}{Numeric vector. Default is 0}

\item{current_velocity}{Numeric vector. Default is 0}

\item{mass}{Numeric vector. Default is 75kg}

\item{weight}{Numeric vector. Default is mass x 9.81}

\item{push_off_distance}{Numeric vector. Default is 0.4}

\item{max_force}{Numeric value. Default is 3000N}

\item{max_velocity}{Numeric vector. Default is 4m/s}

\item{start_perc}{Numeric vector. Default is 0.8}

\item{threshold}{Numeric vector. Default is 0.9}

\item{time_to_max_activation}{Numeric vector. Default is 0.3s}
}
\value{
Data frame with 21 columns
}
\description{
\code{fgen_get_output} returns acceleration and force data based on the current system state (\code{current_time},
    \code{current_distance}, and \code{current_velocity}), system constraints (\code{mass}, \code{weight}, and \code{push_off_distance}),
    and parameters of the Forge Generator (\code{max_force}, \code{current_velocity}, \code{start_perc}, \code{threshold},
    and \code{time_to_max_activation})
#
}
\examples{
state <- expand.grid(
current_time = seq(0, 0.25, length.out = 1000),
current_distance = c(0.1, 0.15, 0.2, 0.25, 0.3),
current_velocity = 0)

fgen_data <- fgen_get_output(
  max_force = 3000,
  mass = 75,
  current_time = state$current_time,
  current_distance = state$current_distance,
  current_velocity = state$current_velocity
)
}
