% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_generator_functions.R
\name{fgen_get_output}
\alias{fgen_get_output}
\title{Get instanteous output of the Force Generator}
\usage{
fgen_get_output(
  mass = 75,
  current_time = 0,
  current_distance = 0,
  current_velocity = 0,
  max_force = 3000,
  gravity_const = 9.81,
  ...
)
}
\arguments{
\item{mass}{Numeric value}

\item{current_time}{Numeric value}

\item{current_distance}{Numeric value}

\item{current_velocity}{Numeric value}

\item{max_force}{Numeric value}

\item{gravity_const}{Numeric value. Default is 9.81}

\item{...}{Use to forward parameters to \code{\link{fgen_force_length}},
\code{\link{fgen_force_time}}, and \code{\link{fgen_force_velocity}} functions}
}
\value{
Data frame
}
\description{
\code{fgen_get_output} returns acceleration and force data based on the current state
    of the system (\code{current_time}, mass, \code{current_velocity}, \code{current_distance}). Parameters in the \code{...} are forwarded to
    \code{\link{fgen_force_length}}, \code{\link{fgen_force_time}}, and \code{\link{fgen_force_velocity}}
    functions that make Force Generator characteristics.
}
\examples{
state <- expand.grid(
current_time = seq(0, 0.25, length.out = 1000),
current_distance = c(0.1, 0.15, 0.2, 0.25, 0.3),
current_velocity = 0)

fgen_data <- fgen_get_output(
  max_force = 3000,
  mass = 75,
  current_time = state$current_time,
  current_distance = state$current_distance,
  current_velocity = state$current_velocity
)
}
