% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_data.R
\docType{data}
\name{testing_data}
\alias{testing_data}
\title{Example testing data}
\format{An object of class \code{data.frame} with 24 rows and 5 columns.}
\usage{
data(testing_data)
}
\description{
Profiling data for N=5 athletes using progressive squat jump measured using jump mat.
    Jump mat has a mesurement error, with systematic bias of 0, proportional bias of 1
    and random error normaly distributed with SD of 0.01s
}
\details{
\describe{
   \item{athlete}{Character string}
   \item{bodyweight}{Bodyweight in kilograms}
   \item{push_off_distance}{Push-off distance during squat jump in meters}
   \iterm{external_load}{External load added to bodyweight in kilograms}
   \item{aerial_time}{Aerial time measured with jump mat in seconds}
}
}
\keyword{datasets}
