% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vj_simulate}
\alias{vj_simulate}
\title{Vertical Jump Simulation}
\usage{
vj_simulate(
  mass = 75,
  weight = mass * gravity_const,
  push_off_distance = 0.4,
  gravity_const = 9.81,
  time_step = 0.01,
  save_trace = TRUE,
  fgen_func = fgen_get_output,
  iter = FALSE,
  max_iter = 1000,
  ...
)
}
\arguments{
\item{gravity_const}{Numeric value. Default is 9.81}

\item{max_iter}{Numeric value. Defaut value 1000}

\item{...}{Forwarded to \code{fgen_func}}
}
\value{

}
\description{
Vertical Jump Simulation
}
\examples{

}
