% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vjsim_data.R
\docType{data}
\name{vjsim_data}
\alias{vjsim_data}
\title{vjsim data}
\format{
A data frame with 8640 rows and 194 variables:
\describe{
   \item{force_generator.bodyweight}{}
   \item{force_generator.push_off_distance}{}
   \item{force_generator.max_force}{}
   \item{force_generator.max_velocity}{}
   \item{force_generator.decline_rate}{}
   \item{force_generator.peak_location}{}
   \item{force_generator.time_to_max_activation}{}
   \item{force_generator.F0}{}
   \item{force_generator.F0_rel}{}
   \item{force_generator.V0}{}
   \item{force_generator.Pmax}{}
   \item{force_generator.Pmax_rel}{}
   \item{force_generator.Sfv}{}
   \item{force_generator.Sfv_rel}{}
   \item{force_generator.take_off_velocity}{}
   \item{force_generator.height}{}
   \item{force_generator.optimal_F0}{}
   \item{force_generator.optimal_F0_rel}{}
   \item{force_generator.optimal_V0}{}
   \item{force_generator.optimal_height}{}
   \item{force_generator.optimal_height_diff}{}
   \item{force_generator.optimal_height_ratio}{}
   \item{force_generator.optimal_Pmax}{}
   \item{force_generator.optimal_Pmax_rel}{}
   \item{force_generator.optimal_take_off_velocity}{}
   \item{force_generator.optimal_take_off_velocity_diff}{}
   \item{force_generator.optimal_take_off_velocity_ratio}{}
   \item{force_generator.optimal_Sfv}{}
   \item{force_generator.optimal_Sfv_rel}{}
   \item{force_generator.Sfv_perc}{}
   \item{force_generator.FV_imbalance}{}
   \item{force_generator.probe_IMB}{}
   \item{bodyweight_jump.take_off_time}{}
   \item{bodyweight_jump.take_off_velocity}{}
   \item{bodyweight_jump.height}{}
   \item{bodyweight_jump.mean_GRF_over_distance}{}
   \item{bodyweight_jump.mean_GRF_over_time}{}
   \item{bodyweight_jump.mean_velocity}{}
   \item{bodyweight_jump.work_done}{}
   \item{bodyweight_jump.impulse}{}
   \item{bodyweight_jump.mean_power}{}
   \item{bodyweight_jump.mean_RFD}{}
   \item{bodyweight_jump.mean_RPD}{}
   \item{bodyweight_jump.peak_GRF}{}
   \item{bodyweight_jump.peak_GRF_time}{}
   \item{bodyweight_jump.peak_GRF_distance}{}
   \item{bodyweight_jump.peak_velocity}{}
   \item{bodyweight_jump.peak_velocity_time}{}
   \item{bodyweight_jump.peak_velocity_distance}{}
   \item{bodyweight_jump.peak_power}{}
   \item{bodyweight_jump.peak_power_distance}{}
   \item{bodyweight_jump.peak_power_time}{}
   \item{bodyweight_jump.peak_RFD}{}
   \item{bodyweight_jump.peak_RFD_distance}{}
   \item{bodyweight_jump.peak_RFD_time}{}
   \item{bodyweight_jump.GRF_at_100ms}{}
   \item{bodyweight_jump.GRF_at_200ms}{}
   \item{bodyweight_jump.peak_RPD}{}
   \item{bodyweight_jump.peak_RPD_distance}{}
   \item{bodyweight_jump.peak_RPD_time}{}
   \item{bodyweight_jump.diff_peak_to_take_off_velocity}{}
   \item{bodyweight_jump.ratio_peak_to_take_off_velocity}{}
   \item{bodyweight_jump.diff_mean_to_take_off_velocity}{}
   \item{bodyweight_jump.ratio_mean_to_take_off_velocity}{}
   \item{bodyweight_jump.diff_mean_to_peak_velocity}{}
   \item{bodyweight_jump.ratio_mean_to_peak_velocity}{}
   \item{bodyweight_jump.diff_mean_to_peak_power}{}
   \item{bodyweight_jump.ratio_mean_to_peak_power}{}
   \item{bodyweight_jump.diff_mean_GRF_over_distance_to_time}{}
   \item{bodyweight_jump.ratio_mean_GRF_over_distance_to_time}{}
   \item{bodyweight_jump.mean_velocity_as_TOV_half}{}
   \item{probe_bodyweight_jump.force_height}{}
   \item{probe_bodyweight_jump.force_height_diff}{}
   \item{probe_bodyweight_jump.force_height_ratio}{}
   \item{probe_bodyweight_jump.velocity_height}{}
   \item{probe_bodyweight_jump.velocity_height_diff}{}
   \item{probe_bodyweight_jump.velocity_height_ratio}{}
   \item{probe_bodyweight_jump.activation_height}{}
   \item{probe_bodyweight_jump.activation_height_diff}{}
   \item{probe_bodyweight_jump.activation_height_ratio}{}
   \item{probe_bodyweight_jump.velocity_to_force_height_diff}{}
   \item{probe_bodyweight_jump.velocity_to_force_height_ratio}{}
   \item{probe_bodyweight_jump.velocity_to_activation_height_diff}{}
   \item{probe_bodyweight_jump.velocity_to_activation_height_ratio}{}
   \item{probe_bodyweight_jump.force_to_velocity_height_diff}{}
   \item{probe_bodyweight_jump.force_to_velocity_height_ratio}{}
   \item{probe_bodyweight_jump.force_to_activation_height_diff}{}
   \item{probe_bodyweight_jump.force_to_activation_height_ratio}{}
   \item{probe_bodyweight_jump.activation_to_velocity_height_diff}{}
   \item{probe_bodyweight_jump.activation_to_velocity_height_ratio}{}
   \item{probe_bodyweight_jump.activation_to_force_height_diff}{}
   \item{probe_bodyweight_jump.activation_to_force_height_ratio}{}
   \item{bosco.height_2BW}{}
   \item{bosco.index}{}
   \item{profile_mean_FV.F0}{}
   \item{profile_mean_FV.F0_rel}{}
   \item{profile_mean_FV.V0}{}
   \item{profile_mean_FV.Pmax}{}
   \item{profile_mean_FV.Pmax_rel}{}
   \item{profile_mean_FV.Sfv}{}
   \item{profile_mean_FV.Sfv_rel}{}
   \item{profile_mean_power.Pmax}{}
   \item{profile_mean_power.Pmax_rel}{}
   \item{profile_mean_power.Pmax_location}{}
   \item{profile_mean_power.F0_perc}{}
   \item{profile_peak_FV.F0}{}
   \item{profile_peak_FV.F0_rel}{}
   \item{profile_peak_FV.V0}{}
   \item{profile_peak_FV.Pmax}{}
   \item{profile_peak_FV.Pmax_rel}{}
   \item{profile_peak_FV.Sfv}{}
   \item{profile_peak_FV.Sfv_rel}{}
   \item{profile_peak_power.Pmax}{}
   \item{profile_peak_power.Pmax_rel}{}
   \item{profile_peak_power.Pmax_location}{}
   \item{profile_peak_power.F0_perc}{}
   \item{profile_load_take_off_velocity.V0}{}
   \item{profile_load_take_off_velocity.L0}{}
   \item{profile_load_take_off_velocity.L0_rel}{}
   \item{profile_load_take_off_velocity.Imax}{}
   \item{profile_load_take_off_velocity.Imax_rel}{}
   \item{profile_load_take_off_velocity.Slv}{}
   \item{profile_load_take_off_velocity.Slv_rel}{}
   \item{profile_load_impulse.Imax}{}
   \item{profile_load_impulse.Imax_rel}{}
   \item{profile_load_impulse.Imax_location}{}
   \item{profile_load_impulse.L0_perc}{}
   \item{LPF_profile.slope}{}
   \item{LPF_profile.slope_rel}{}
   \item{samozino_theoretical_profile.F0}{}
   \item{samozino_theoretical_profile.F0_rel}{}
   \item{samozino_theoretical_profile.V0}{}
   \item{samozino_theoretical_profile.Pmax}{}
   \item{samozino_theoretical_profile.Pmax_rel}{}
   \item{samozino_theoretical_profile.Sfv}{}
   \item{samozino_theoretical_profile.Sfv_rel}{}
   \item{samozino_theoretical_profile.take_off_velocity}{}
   \item{samozino_theoretical_profile.height}{}
   \item{samozino_theoretical_profile.optimal_F0}{}
   \item{samozino_theoretical_profile.optimal_F0_rel}{}
   \item{samozino_theoretical_profile.optimal_V0}{}
   \item{samozino_theoretical_profile.optimal_height}{}
   \item{samozino_theoretical_profile.optimal_height_diff}{}
   \item{samozino_theoretical_profile.optimal_height_ratio}{}
   \item{samozino_theoretical_profile.optimal_Pmax}{}
   \item{samozino_theoretical_profile.optimal_Pmax_rel}{}
   \item{samozino_theoretical_profile.optimal_take_off_velocity}{}
   \item{samozino_theoretical_profile.optimal_take_off_velocity_diff}{}
   \item{samozino_theoretical_profile.optimal_take_off_velocity_ratio}{}
   \item{samozino_theoretical_profile.optimal_Sfv}{}
   \item{samozino_theoretical_profile.optimal_Sfv_rel}{}
   \item{samozino_theoretical_profile.Sfv_perc}{}
   \item{samozino_theoretical_profile.FV_imbalance}{}
   \item{samozino_theoretical_profile.probe_IMB}{}
   \item{samozino_practical_profile.F0}{}
   \item{samozino_practical_profile.F0_rel}{}
   \item{samozino_practical_profile.V0}{}
   \item{samozino_practical_profile.Pmax}{}
   \item{samozino_practical_profile.Pmax_rel}{}
   \item{samozino_practical_profile.Sfv}{}
   \item{samozino_practical_profile.Sfv_rel}{}
   \item{samozino_practical_profile.take_off_velocity}{}
   \item{samozino_practical_profile.height}{}
   \item{samozino_practical_profile.optimal_F0}{}
   \item{samozino_practical_profile.optimal_F0_rel}{}
   \item{samozino_practical_profile.optimal_V0}{}
   \item{samozino_practical_profile.optimal_height}{}
   \item{samozino_practical_profile.optimal_height_diff}{}
   \item{samozino_practical_profile.optimal_height_ratio}{}
   \item{samozino_practical_profile.optimal_Pmax}{}
   \item{samozino_practical_profile.optimal_Pmax_rel}{}
   \item{samozino_practical_profile.optimal_take_off_velocity}{}
   \item{samozino_practical_profile.optimal_take_off_velocity_diff}{}
   \item{samozino_practical_profile.optimal_take_off_velocity_ratio}{}
   \item{samozino_practical_profile.optimal_Sfv}{}
   \item{samozino_practical_profile.optimal_Sfv_rel}{}
   \item{samozino_practical_profile.Sfv_perc}{}
   \item{samozino_practical_profile.FV_imbalance}{}
   \item{samozino_practical_profile.probe_IMB}{}
   \item{simple_profile.L0}{}
   \item{simple_profile.TOV0}{}
   \item{simple_profile.Sfv}{}
   \item{simple_profile.Sfv_rel}{}
   \item{simple_profile.take_off_velocity}{}
   \item{simple_profile.height}{}
   \item{simple_profile.optimal_L0}{}
   \item{simple_profile.optimal_TOV0}{}
   \item{simple_profile.optimal_Sfv}{}
   \item{simple_profile.optimal_Sfv_rel}{}
   \item{simple_profile.optimal_take_off_velocity}{}
   \item{simple_profile.optimal_height}{}
   \item{simple_profile.Sfv_perc}{}
   \item{simple_profile.FV_imbalance}{}
   \item{simple_profile.probe_IMB}{}
}
}
\usage{
data(vjsim_data)
}
\description{
Data generated through simulation. Used for exploring relationship between Force Generator
characteristics as latent or intrinsic qualities and manifested performance in a single jump
as well as profile metrics and optimization metrics
}
\keyword{datasets}
